requirements.txt
customtkinter
pvporcupine
SpeechRecognition
huggingface_hub
requests
pyaudio
pyinstaller
chromadb
sentence-transformers
PyPDF2
python-docx
BeautifulSoup4
pyspellchecker
g4f
langdetect
deep_translator
edge-tts
playsound
rarfile
spacy